requirements.txt
actionlib==1.14.3
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
bcrypt==3.1.7
bondpy==1.8.7
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.1.1
cryptography==2.8
cv-bridge==1.13.0.post0
cycler==0.12.1
defusedxml==0.6.0
distro==1.9.0
docutils==0.16
dynamic-reconfigure==1.7.6
empy==3.3.2
exceptiongroup==1.3.0
filelock==3.16.1
fiona==1.10.1
fonttools==4.57.0
fsspec==2025.3.0
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
geopandas==0.13.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
jiter==0.9.1
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.17.4
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==1.106.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
paramiko==2.6.0
pillow==10.4.0
pycryptodomex==3.6.1
pydantic==2.10.6
pydantic-core==2.27.2
PyNaCl==1.3.0
pyparsing==2.4.6
pyproj==3.5.0
python-astar==0.1.0
python-dateutil==2.9.0.post0
python-gnupg==0.4.5
pytz==2025.2
PyYAML==5.3.1
regex==2024.11.6
requests==2.32.4
roman==2.0.0
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.4
rosinstall==0.7.8
roslaunch==1.17.4
roslib==1.15.10
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sensor-msgs==1.13.2
shapely==2.0.7
sip==4.19.21
six==1.14.0
smclib==1.8.7
sniffio==1.3.1
sympy==1.13.3
tf==1.13.4
tf-conversions==1.13.4
tf2-py==0.7.10
tf2-ros==0.7.10
threadpoolctl==3.5.0
tokenizers==0.20.3
topic-tools==1.17.4
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
tsp-solver2==0.4.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vcstools==0.1.42
wstool==0.1.18
zipp==3.20.2
